# internal/data/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.20.0

package db

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Profile struct {
	ID        string
	Name      string
	AvatarUrl pgtype.Text
	FirstName pgtype.Text
	LastName  pgtype.Text
	UpdatedAt pgtype.Timestamptz
	CreatedAt pgtype.Timestamptz
}

type Section struct {
	ID           pgtype.UUID
	TechnologyID pgtype.UUID
	Title        string
	Description  pgtype.Text
	ImageUrl     pgtype.Text
	Position     int32
	UpdatedAt    pgtype.Timestamptz
	CreatedAt    pgtype.Timestamptz
}

type Task struct {
	ID          pgtype.UUID
	SectionID   pgtype.UUID
	Title       string
	Description pgtype.Text
	ImageUrl    pgtype.Text
	Difficulty  int32
	Content     []byte
	Position    pgtype.Int4
	IsPublic    bool
	UpdatedAt   pgtype.Timestamptz
	CreatedAt   pgtype.Timestamptz
}

type Technology struct {
	ID          pgtype.UUID
	Title       string
	Description pgtype.Text
	ImageUrl    pgtype.Text
	Position    int32
	UpdatedAt   pgtype.Timestamptz
	CreatedAt   pgtype.Timestamptz
}